+++ examples/xApp/c/orange/xapp_es_with_cell_util.c
@@ -35,6 +35,7 @@
 #include "../../../../src/sm/rc_sm/rc_sm_id.h"
 #include "../../../../src/sm/rc_sm/ie/rc_data_ie.h"
 #include "../../../../src/util/e.h"
+#include <unistd.h>
 
 #define MIN_SINR -10
 
@@ -389,7 +390,7 @@ void add_neighCell(struct SINRServingValues* UE, const uint16_t neighCellID, con
   }
 }
 
-uint8_t getTargetCellID(callback_data_t data) 
+uint16_t getTargetCellID(callback_data_t data)
 {
   assert(data.neighCells != NULL);
 
@@ -1308,7 +1309,7 @@ uint16_t doHandoverAction(callback_data_t data)
   return handover_sent ? 1 : 0;
 }
 
-uint8_t switchOffCurrentCell(callback_data_t data) 
+uint16_t switchOffCurrentCell(callback_data_t data)
 {
   rc_ctrl_req_data_t rc_ctrl = {0};
   ue_id_e2sm_t ue_id = gen_rc_ue_id(GNB_UE_ID_E2SM, data.ueID);
